Typescript support for docusaurus - started writing documentation for Do.Try
