--- documentation/docs/doc2.md
@@ -1,6 +1,0 @@
----
-id: doc2
-title: Document Number 2
----
-
-This is a link to [another document.](doc3.md) This is a link to an [external page.](http://www.example.com/)
