--- documentation/docs/interactiveDoc.mdx
@@ -1,27 +1,0 @@
----
-id: mdx
-title: Powered by MDX
----
-
-You can write JSX and use React components within your Markdown thanks to [MDX](https://mdxjs.com/).
-
-The `.mdx` extension is not required, but will enable better support from tooling (IDE, Prettier...).
-
-export const Highlight = ({children, color}) => (
-  <span
-    style={{
-      backgroundColor: color,
-      borderRadius: '2px',
-      color: '#fff',
-      padding: '0.2rem',
-    }}
-    onClick={() => alert('Highlight pressed!')}>
-    {children}
-  </span>
-);
-
-<Highlight color="#25c2a0">Docusaurus green</Highlight> and <Highlight color="#1877F2">
-  Facebook blue
-</Highlight> are my favorite colors.
-
-I can write **Markdown** alongside my _JSX_!
