--- documentation/sidebars.js
@@ -1,6 +1,0 @@
-module.exports = {
-  someSidebar: {
-    Docusaurus: ['doc1', 'doc2', 'doc3'],
-    Features: ['mdx'],
-  },
-};
