+++ documentation/sidebars.ts
@@ -1,0 +1,5 @@
+module.exports = {
+  primary: {
+    Packages: ['index'],
+  },
+};
