+++ documentation/tsconfig.json
@@ -1,0 +1,4 @@
+{
+    "extends": "@tsconfig/docusaurus/tsconfig.json",
+    "include": ["src/"]
+}
